refactor: renamed the folder of application & project related functionality, making it more intuitive
